refactor: ContextualHelpSection 컴포넌트를 작은 컴포넌트들로 분리
- useDragAndDrop, usePositioning, useHelpContent 커스텀 훅 추출
- HelpContentDisplay, VideoSection, DragHandle 하위 컴포넌트 생성
- 메인 컴포넌트를 532줄에서 ~170줄로 축소
- 동일한 기능과 API 유지
- 데스크톱 도움말 섹션 확장 클릭 동작 수정
